Move date module out of detail namespace
